*** empty log message ***

+++ v3.x/user/msg_clock/cmdcrn.c
@@ -35,7 +35,7 @@ void CommandChrono_Put(char command)
 {
 	static RT_TASK *ackn = 0;
 	unsigned int put = 'c';
-	unsigned int msg;
+	unsigned long msg;
 
 	if (ackn != rt_get_adr(nam2num("CHRTSK"))) {
 		ackn = rt_rpc(rt_get_adr(nam2num("CHRTSK")), put, &msg);
